refactor: make examples single file
