+++ examples/simple-static-router/README.md
@@ -1,0 +1,1 @@
+# Simple Static Router
