+++ examples/simple-static-router/main.go
